[idp saml2] make authncontext generation customizable (by session, or by authentication event)
In the use user backend case, we lookup the backend used to authenticate
in the session and generate a corresponding authn contextx. In the other
case if no authentication events exist for the current AuthnRequest, we
assume PREVIOUS_SESSION as the AuthnContext (PREVIOUS_SESSION_ON_SECURE_TRANSPORT
would be better but it does not exist).