[authsaml2] extract attributes from assertions, compact values as string containing XML content
There is still some loss of content because saml:AttributeValue can hold
attributes from any third party namespaces for which we do not have a
way to transmit the value currently. The most frequent being xsi:type to
transmit a type to interpret the contained string value of the attribute
(for example base64Binary to transmit binary blobs as AttributeValue).

Without supporting all third party namespace attribute, we could support
the most frequent xsd:type values, and convert the attributevalue to a
python object as needed.