[authsaml2] make authentication the django way
There is now two backends one for transient and another for persistent
authentications. The first one never create an object in the db. the
second one do as before. It properly uses
django.auth.contrib.authenticate() so that if we want to implement an
LDAP/SAML backend, we can.

The next step would be to authenticate based on an assertion, not just
the NameID. In the transient case it would open the way to storing
attributes into the session, the created user id would contain a copy of
the session key so that we can retrieve the attributes in the backend
get_user() method (which do not receive the request, so it does not have
a direct access to the session).