[saml track provider sources 2/3: add a --source option to the load-metadata management command
If --source is used, all new providers are created with this source
value, and all provider with the same source but not present in the
metadata files are removed.

If used with --delete, after removing all the providers whose entity id
was present in the metadata files, we remove all the providers with this
source.