[idp] Add attribute aggregator application
Intro
-----

Attribute aggregator provides a main Model class called UserAttributeProfile,
functions to load attributes and extract attributes.

The mapping between attribute namespaces is built-in and depends on a unique
file (mapping.py).

A main schema is defined and is based on LDAP/X500 for naming. The support
of http://schemas.xmlsoap.org/ws/2005/05/identity/claims is partly complete.

Source of attributes are connected with attribute loading functions using
signals.

FAQ
---

Why not use the Django User profile?

The django user profile needs to define attributes as class attributes and then
support form filling or mapping with LDAP.

That is useful and may be used, especially because the profile can be used as a
source of attribute to load in the attribute_aggregator profile.

The attribute_aggregator profile allow to load multivalued attributes from any
source supported (LDAP, Django profile and attributes in Django session for now)
from any namespace defined in mapping.py (LDAP/X500 and claims for now).

The profile can be loaded giving a source or a list of attribute, or can be
from any known source, or with a dictionnary.

Attributes can be extracted with many functions in any namespace supported.

Quick explanation
-----------------

The schema is defined in mapping and is made of definitions like this:

"sn": {
    "oid": "2.5.4.4",
    "display_name": _("sn surname"),
    "alias": ['surname'],
    "profile_field_name": 'last_name',
    "type": "http://www.w3.org/2001/XMLSchema#string",
    "namespaces": {
        "http://schemas.xmlsoap.org/ws/2005/05/identity/claims": {
            "identifiers":
                [
            "http://schemas.xmlsoap.org/ws/2005/05/identity/claims/surname",
                ],
            "friendly_names":
                [
            "Last Name",
                ],
        }
    }
},

The profile store all the data in a text field taht contains a cPickle list of
instances of the class AttributeData.

The profile is attached to a user and then can be created or loaded with

    profile = load_or_create_user_profile(user=user)

User may be None to create a temporary profile for an anonymous user. But
that need a DB cleaning function not implemented.

The AttributeData instances have a field expiration_data. It the profile exists,
obsolete data are removed at loading.