[idp] New model for attribute policy for pushing attributes to SP
The AttributePolicy class aims to associate instance to service providers
to indicate which attributes must be pushed in the response.

The options include where attributes must be taken from synchronous source or
from assynchronous source, considered with attributes in session.

The options also indicates the attribute namespace and name format to employ
in the response.