[idp] Support profile loading and extraction according to sp policy
The function receive as parameter a service provider name and a user
instance.

If there is no service provider name, no policy can be found, thus
the function return no attributes.

If the user is None, attributes in request.session may be used.

The function analyse the policy then load the profile. Next, the
function extract attributes from the profile and create a dictionnary
attributes with the attribute name and name format according to the policy.