[authsaml2] Load attributes in session once again.
To keep compatibility with univnautes that use the attribute dictionnary
in session.

The attributes are loaded in another dictionnary with name:

  request.session['multisource_attributes'].

This is necessary because we assume that multiple bindings
can load attributes in the session during the same user session.
Then, all attributes are loaded in this dictionnary indicating the
source name:

  request.session['multisource_attributes'] \
        [source_name] = dict()