[saml][authsaml2] Modify SP configuration and add SP SAML2 options policies
 * Model modificiations
   * MyServiceProvider deleted
   * IdPOptionsSPPolicy modified
   * SPOptionsIdPPolicy created
   * LibertyServiceProvider modified
 * Script modification
   * sync-metadata

 - The model MyServiceProvider used to configure settings related to the SAML2
service provider role has been removed. These settings has been included in
the model IdPOptionsSPPolicy. Then, these settings can now be configured per
IdP.

 - The SPOptionsIdPPolicy model has been created to manage policies of
service provider SAML2 options. Like IdPOptionsSPPolicy policies, they are
managed with the concept of 'All', associated with a LibertyServiceProvider
instance and 'Default'

    Reminder
    --------

    The policy management with global policies is nearly used for any kind of
    policy in Authentic2.

    For each kind of these policies, the system takes in account two special
    global policies named 'Default' and 'All':
       * If no other policy applies, the policy 'Default' will apply.
       * A policy can be created and attached to any related object. This
         policy is authoritative on policy 'Default'.
       * If the policy 'All' exists, it is authoritative on any
         other policy.
       * The global policies must be created by the administrator if
         necessary.

     **A policy is taken in account only if it is enabled.**

 - The script sync-metadata has been updated because the SAML2 options are not
set directly on LibertyProviders anymore. The script options
--sp-default-nameid-format and --sp-accepted-nameid-format have been removed.
The allow_create = True parameter for identity providers has also been
removed.

To configure the SAML2 parameters of providers imported with the script,
either a global policy, 'Default' or 'All' is defined in the administration
interface, or a regular policy is created and its name is given in parameter
of the script. A policy name for identity providers is given with the option
--idp-policy and with the option --sp-policy for the service providers.